+++ Changelog.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[//]: # (## [Unreleased])
+
+## [2.2.0] - 2022-04-04
+
+### Added
+- Scriptable Object can now be added to a Prefab or another Scriptable Object.
+- Samples to demonstrate adding Scriptable Objects to a Prefab and other Scriptable Object.
+- Refrenece : https://tsubakit1.hateblo.jp/entry/2017/08/03/233000
 
 ## [2.1.4] - 2022-03-29
 
@@ -97,7 +104,8 @@ Major version upgrade. Not backward compatible with version 1.0.x.
 - Raise Button on Inspector. Click the button to simulate GameEvent Raise from inspector.
 - CancellationTokenUtility. Helper class for cancelling and refreshing CancellationToken.
 
-[Unreleased]: https://github.com/kadinche/Kassets/compare/2.1.4...HEAD
+[Unreleased]: https://github.com/kadinche/Kassets/compare/2.2.0...HEAD
+[2.2.0]: https://github.com/kadinche/Kassets/compare/2.1.4...2.2.0
 [2.1.4]: https://github.com/kadinche/Kassets/compare/2.1.3...2.1.4
 [2.1.3]: https://github.com/kadinche/Kassets/compare/2.1.2...2.1.3
 [2.1.2]: https://github.com/kadinche/Kassets/compare/2.1.1...2.1.2